ISSUE #4321 - Fixed circular dependency in viewpoint selector
