+++ frontend/src/v4/modules/viewpoints/viewpoints.sagas.ts
@@ -23,15 +23,12 @@ import { isV5 } from '@/v4/helpers/isV5';
 
 import { prefixBaseDomain } from '@/v5/helpers/url.helper';
 import { getAPIUrl } from '@/v4/services/api/default';
-import { CHAT_CHANNELS } from '../../constants/chat';
 import { ROUTES } from '../../constants/routes';
 import { prepareGroup } from '../../helpers/groups';
-import { createGroupsFromViewpoint, generateViewpoint, groupsOfViewpoint,
-	isViewpointLoaded,
+import { createGroupsFromViewpoint, generateViewpoint,
 	mergeGroupsDataFromViewpoint, setGroupData } from '../../helpers/viewpoints';
 import * as API from '../../services/api';
 import { Viewer } from '../../services/viewer/viewer';
-import { ChatActions } from '../chat';
 import { DialogActions } from '../dialog';
 import { GroupsActions } from '../groups';
 import { IssuesActions } from '../issues';
@@ -46,7 +43,7 @@ import { waitForTreeToBeReady } from '../tree/tree.sagas';
 import { ViewerGuiActions } from '../viewerGui';
 import { PRESET_VIEW } from './viewpoints.constants';
 import { ViewpointsActions, ViewpointsTypes } from './viewpoints.redux';
-import { selectSelectedViewpoint, selectViewpointsGroups, selectViewpointsGroupsBeingLoaded } from '.';
+import { groupsOfViewpoint, isViewpointLoaded, selectSelectedViewpoint, selectViewpointsGroups, selectViewpointsGroupsBeingLoaded } from '.';
 
 export const getThumbnailUrl = (thumbnail) => getAPIUrl(thumbnail);
